+++ worlds/banjo_tooie/WorldOrder.py
@@ -14,30 +14,17 @@ else:
 # Shamelessly Stolen from KH2 :D
 
 
-def WorldRandomize(world: BanjoTooieWorld) -> None:
-    # Universal Tracker Magic
-    if hasattr(world.multiworld, "re_gen_passthrough"):
-        if "Banjo-Tooie" in world.multiworld.re_gen_passthrough:
-            passthrough = world.multiworld.re_gen_passthrough["Banjo-Tooie"]
-            world.world_requirements = passthrough["world_order"]
-            world.world_order = passthrough["world_keys"]
-            world.worlds_randomized = bool(passthrough["worlds"] == "true")
-            world.starting_egg = passthrough["starting_egg"]
-            world.starting_attack = passthrough["starting_attack"]
-            world.jamjars_siloname_costs = passthrough["jamjars_siloname_costs"]
-            world.loading_zones = passthrough["loading_zones"]
-    else:
-        randomize_level_order(world)
-        set_level_costs(world)
-        randomize_entrance_loading_zones(world)
-        randomize_boss_loading_zones(world)
-        choose_unlocked_silos(world)
-        handle_early_moves(world)
-        generate_jamjars_costs(world)
+def randomize_world_progression(world: BanjoTooieWorld) -> None:
+    randomize_level_order(world)
+    set_level_costs(world)
+    randomize_entrance_loading_zones(world)
+    randomize_boss_loading_zones(world)
+    choose_unlocked_silos(world)
+    handle_early_moves(world)
+    generate_jamjars_costs(world)
 
 def randomize_level_order(world: BanjoTooieWorld) -> None:
-    world.worlds_randomized = world.options.randomize_worlds
-    if not world.worlds_randomized:
+    if not world.options.randomize_worlds:
         world.world_order = {
             regionName.MT:  1230944, #These ids stay in the same order, but the keys may switch order when randomized.
             regionName.GM:  1230945,